Extract ColorPickerUI from ColorpickerPreference
